+++ CHANGELOG.md
@@ -7,11 +7,11 @@
 ### First Release
 
 #### Module ryan_joiner
+- Added ``_statistic`` function (06/11/2023);
 - Added ``_p_value`` function (06/11/2023);
 - Added ``_order_statistic`` function (06/11/2023);
 - Added ``_normal_order_statistic`` function (05/11/2023);
 - Added ``_critical_value`` function (06/11/2023);
-- Added ``statistic`` function (05/11/2023);
 - Added ``citation`` function (04/11/2023);